lesson 8/requirements.txt
mcp
fastmcp
crewai
crewai-tools[mcp]
langchain-openai
langchain-core
python-dotenv
json

